Make some services thread-safe

+++ Energize/Services/Senders/PaginatorSenderService.cs
@@ -5,6 +5,7 @@ using Energize.Essentials.MessageConstructs;
 using Energize.Interfaces.Services.Listeners;
 using Energize.Interfaces.Services.Senders;
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
@@ -16,7 +17,7 @@ namespace Energize.Services.Senders
     [Service("Paginator")]
     public class PaginatorSenderService : ServiceImplementationBase, IPaginatorSenderService
     {
-        private readonly Dictionary<ulong, Paginator<object>> Paginators;
+        private readonly ConcurrentDictionary<ulong, Paginator<object>> Paginators;
         private readonly MessageSender MessageSender;
         private readonly ServiceManager ServiceManager;
         private readonly Logger Logger;
@@ -24,7 +25,7 @@ namespace Energize.Services.Senders
 
         public PaginatorSenderService(EnergizeClient client)
         {
-            this.Paginators = new Dictionary<ulong, Paginator<object>>();
+            this.Paginators = new ConcurrentDictionary<ulong, Paginator<object>>();
             this.MessageSender = client.MessageSender;
             this.ServiceManager = client.ServiceManager;
             this.Logger = client.Logger;
@@ -38,7 +39,7 @@ namespace Energize.Services.Senders
                         toRemove.Add(paginator.Key);
 
                 foreach (ulong msgId in toRemove)
-                    this.Paginators.Remove(msgId);
+                    this.Paginators.TryRemove(msgId, out Paginator<object> _);
                 
                 if(toRemove.Count > 0)
                     this.Logger.Nice("Paginator", ConsoleColor.Gray, $"Cleared {toRemove.Count} paginator instance{(toRemove.Count == 1 ? string.Empty : "s")}");
@@ -77,8 +78,8 @@ namespace Energize.Services.Senders
             if (posted != null)
             {
                 paginator.Message = posted;
-                this.Paginators.Add(posted.Id, paginator.ToObject());
-                this.AddReactions(posted, "◀", "⏹", "▶");
+                if (this.Paginators.TryAdd(posted.Id, paginator.ToObject()))
+                    this.AddReactions(posted, "◀", "⏹", "▶");
             }
 
             return posted;
@@ -99,8 +100,8 @@ namespace Energize.Services.Senders
             if (posted != null)
             {
                 paginator.Message = posted;
-                this.Paginators.Add(posted.Id, paginator.ToObject());
-                this.AddReactions(posted, "◀", "⏹", "▶");
+                if (this.Paginators.TryAdd(posted.Id, paginator.ToObject()))
+                    this.AddReactions(posted, "◀", "⏹", "▶");
             }
 
             return posted;
@@ -114,8 +115,8 @@ namespace Energize.Services.Senders
             if (posted != null)
             {
                 paginator.Message = posted;
-                this.Paginators.Add(posted.Id, paginator.ToObject());
-                this.AddReactions(posted, "◀", "⏹", "▶");
+                if (this.Paginators.TryAdd(posted.Id, paginator.ToObject()))
+                    this.AddReactions(posted, "◀", "⏹", "▶");
             }
 
             return posted;
@@ -129,8 +130,8 @@ namespace Energize.Services.Senders
             if (posted != null)
             {
                 paginator.Message = posted;
-                this.Paginators.Add(posted.Id, paginator.ToObject());
-                this.AddReactions(posted, "◀", "⏹", "⏯", "▶");
+                if (this.Paginators.TryAdd(posted.Id, paginator.ToObject()))
+                    this.AddReactions(posted, "◀", "⏹", "⏯", "▶");
             }
 
             return posted;
@@ -143,8 +144,8 @@ namespace Energize.Services.Senders
             ["▶"] = async (sender, paginator, cache, chan, reaction) => await paginator.Next(),
             ["⏹"] = async (sender, paginator, cache, chan, reaction) =>
             {
-                sender.Paginators.Remove(cache.Value.Id);
-                await chan.DeleteMessageAsync(paginator.Message);
+                if (sender.Paginators.TryRemove(cache.Value.Id, out Paginator<object> _))
+                    await chan.DeleteMessageAsync(paginator.Message);
             },
             ["⏯"] = OnPlayReaction,
         };
@@ -159,9 +160,7 @@ namespace Energize.Services.Senders
         private async Task OnReaction(Cacheable<IUserMessage, ulong> cache, ISocketMessageChannel chan, SocketReaction reaction)
         {
             if (!cache.HasValue || !this.IsValidEmote(reaction)) return;
-            if (!this.Paginators.ContainsKey(cache.Value.Id)) return;
-
-            Paginator<object> paginator = this.Paginators[cache.Value.Id];
+            if (!this.Paginators.TryGetValue(cache.Value.Id, out Paginator<object> paginator)) return;
             if (paginator.UserID != reaction.UserId) return;
 
             await ReactionCallbacks[reaction.Emote.Name](this, paginator, cache, chan, reaction);
